Refactor router testing
This attempts to test as much of the functionality close to the file
they're actually defined.
