+++ manifests/router/link_route.pp
@@ -14,10 +14,11 @@ define qpid::router::link_route(
   Enum['in', 'out'] $direction,
   String $prefix = 'queue.',
   Optional[String] $connection = undef,
+  String $config_file = $qpid::router::config_file,
 ){
 
   concat::fragment {"qdrouter+link_route_${name}.conf":
-    target  => $qpid::router::config_file,
+    target  => $config_file,
     content => template('qpid/router/link_route.conf.erb'),
     order   => '04',
   }
